fixed and bubbling working

+++ src/lib/utils.hpp
@@ -5,7 +5,7 @@
 namespace lib {
 
 template < typename T > T max(T a, T b) {
-  return ((u64)a > (u64)b) ? a : b;
+  return (a > b) ? a : b;
 }
 
 template < typename T > b8 is_equal(T a, T b) {
@@ -13,11 +13,11 @@ template < typename T > b8 is_equal(T a, T b) {
 }
 
 template < typename T > b8 is_smaller(T a, T b) {
-  return (u64)a < (u64)b;
+  return a < b;
 }
 
 template < typename T > b8 is_greater(T a, T b) {
-  return (u64)a > (u64)b;
+  return a > b;
 }
 
 } // namespace lib
